rails 28 checkpoint completed

+++ app/controllers/sessions_controller.rb
@@ -16,8 +16,10 @@ class SessionsController < ApplicationController
   end
 
   def destroy
-    destroy_session(current_user)
-    flash[:notice] = "You've been signed out, come back soon!"
-    redirect_to root_path
+    if current_user
+      destroy_session(current_user)
+      flash[:notice] = "You've been signed out, come back soon!"
+      redirect_to root_path
+    end
   end
 end
